Severity Priority Sync script
